cleanedup code changed the importing of sidebar in navbar to cordinator.js
